requirements.txt
click==8.1.5
colorama==0.4.6
numpy==1.25.1
opencv-python==4.8.0.74
Pillow==10.0.0
scipy==1.11.1
segment-anything==1.0
shapely==2.0.1
tqdm==4.65.0
